% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{create_mesh}
\alias{create_mesh}
\title{Creates a new service mesh}
\usage{
create_mesh(clientToken = NULL, meshName)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.}

\item{meshName}{[required] The name to use for the service mesh.}
}
\description{
Creates a new service mesh. A service mesh is a logical boundary for network traffic between the services that reside within it.
}
\details{
After you create your service mesh, you can create virtual nodes, virtual routers, and routes to distribute traffic between the applications in your mesh.
}
\section{Accepted Parameters}{
\preformatted{create_mesh(
  clientToken = "string",
  meshName = "string"
)
}
}

