% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{create_virtual_router}
\alias{create_virtual_router}
\title{Creates a new virtual router within a service mesh}
\usage{
create_virtual_router(clientToken = NULL, meshName, spec,
  virtualRouterName)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.}

\item{meshName}{[required] The name of the service mesh in which to create the virtual router.}

\item{spec}{[required] The virtual router specification to apply.}

\item{virtualRouterName}{[required] The name to use for the virtual router.}
}
\description{
Creates a new virtual router within a service mesh.
}
\details{
Virtual routers handle traffic for one or more service names within your mesh. After you create your virtual router, create and associate routes for your virtual router that direct incoming requests to different virtual nodes.
}
\section{Accepted Parameters}{
\preformatted{create_virtual_router(
  clientToken = "string",
  meshName = "string",
  spec = list(
    serviceNames = list(
      "string"
    )
  ),
  virtualRouterName = "string"
)
}
}

