% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{delete_mesh}
\alias{delete_mesh}
\title{Deletes an existing service mesh}
\usage{
delete_mesh(
  meshName = "string"
)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete.}
}
\description{
Deletes an existing service mesh.
}
\details{
You must delete all resources (routes, virtual routers, virtual nodes) in the service mesh before you can delete the mesh itself.
}
\examples{

}
