% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{delete_route}
\alias{delete_route}
\title{Deletes an existing route}
\usage{
delete_route(
  meshName = "string",
  routeName = "string",
  virtualRouterName = "string"
)
}
\arguments{
\item{meshName}{[required] The name of the service mesh in which to delete the route.}

\item{routeName}{[required] The name of the route to delete.}

\item{virtualRouterName}{[required] The name of the virtual router in which to delete the route.}
}
\description{
Deletes an existing route.
}
\examples{

}
