% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{delete_virtual_router}
\alias{delete_virtual_router}
\title{Deletes an existing virtual router}
\usage{
delete_virtual_router(meshName, virtualRouterName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh in which to delete the virtual router.}

\item{virtualRouterName}{[required] The name of the virtual router to delete.}
}
\description{
Deletes an existing virtual router.
}
\details{
You must delete any routes associated with the virtual router before you can delete the router itself.
}
\section{Accepted Parameters}{
\preformatted{delete_virtual_router(
  meshName = "string",
  virtualRouterName = "string"
)
}
}

