% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{describe_mesh}
\alias{describe_mesh}
\title{Describes an existing service mesh}
\usage{
describe_mesh(meshName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to describe.}
}
\description{
Describes an existing service mesh.
}
\section{Accepted Parameters}{
\preformatted{describe_mesh(
  meshName = "string"
)
}
}

