% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{describe_route}
\alias{describe_route}
\title{Describes an existing route}
\usage{
describe_route(
  meshName = "string",
  routeName = "string",
  virtualRouterName = "string"
)
}
\arguments{
\item{meshName}{[required] The name of the service mesh in which the route resides.}

\item{routeName}{[required] The name of the route to describe.}

\item{virtualRouterName}{[required] The name of the virtual router with which the route is associated.}
}
\description{
Describes an existing route.
}
\examples{

}
