% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{describe_virtual_node}
\alias{describe_virtual_node}
\title{Describes an existing virtual node}
\usage{
describe_virtual_node(meshName, virtualNodeName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh in which the virtual node resides.}

\item{virtualNodeName}{[required] The name of the virtual node to describe.}
}
\description{
Describes an existing virtual node.
}
\section{Accepted Parameters}{
\preformatted{describe_virtual_node(
  meshName = "string",
  virtualNodeName = "string"
)
}
}

