% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{describe_virtual_router}
\alias{describe_virtual_router}
\title{Describes an existing virtual router}
\usage{
describe_virtual_router(meshName, virtualRouterName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh in which the virtual router resides.}

\item{virtualRouterName}{[required] The name of the virtual router to describe.}
}
\description{
Describes an existing virtual router.
}
\section{Accepted Parameters}{
\preformatted{describe_virtual_router(
  meshName = "string",
  virtualRouterName = "string"
)
}
}

