% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{list_virtual_routers}
\alias{list_virtual_routers}
\title{Returns a list of existing virtual routers in a service mesh}
\usage{
list_virtual_routers(
  limit = 123,
  meshName = "string",
  nextToken = "string"
)
}
\arguments{
\item{limit}{The maximum number of mesh results returned by \code{ListVirtualRouters} in paginated output. When this parameter is used, \code{ListVirtualRouters} only returns \code{limit} results in a single page along with a \code{nextToken} response element. The remaining results of the initial request can be seen by sending another \code{ListVirtualRouters} request with the returned \code{nextToken} value. This value can be between 1 and 100. If this parameter is not used, then \code{ListVirtualRouters} returns up to 100 results and a \code{nextToken} value if applicable.}

\item{meshName}{[required] The name of the service mesh in which to list virtual routers.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated \code{ListVirtualRouters} request where \code{limit} was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the \code{nextToken} value.}
}
\description{
Returns a list of existing virtual routers in a service mesh.
}
\examples{

}
