% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_package.R
\docType{package}
\name{paws.appmesh-package}
\alias{paws.appmesh}
\alias{paws.appmesh-package}
\title{paws.appmesh: AWS App Mesh}
\description{
AWS App Mesh is a service mesh based on the Envoy
    proxy that makes it easy to monitor and control containerized
    microservices. App Mesh standardizes how your microservices
    communicate, giving you end-to-end visibility and helping to ensure
    high-availability for your applications.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
