% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appmesh_operations.R
\name{update_virtual_router}
\alias{update_virtual_router}
\title{Updates an existing virtual router in a specified service mesh}
\usage{
update_virtual_router(clientToken = NULL, meshName, spec,
  virtualRouterName)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.}

\item{meshName}{[required] The name of the service mesh in which the virtual router resides.}

\item{spec}{[required] The new virtual router specification to apply. This overwrites the existing data.}

\item{virtualRouterName}{[required] The name of the virtual router to update.}
}
\description{
Updates an existing virtual router in a specified service mesh.
}
\section{Accepted Parameters}{
\preformatted{update_virtual_router(
  clientToken = "string",
  meshName = "string",
  spec = list(
    serviceNames = list(
      "string"
    )
  ),
  virtualRouterName = "string"
)
}
}

